requirements.txt
# ── LLM ──────────────────────────────────────────────────────
openai>=1.30.0

# ── Env management ───────────────────────────────────────────
python-dotenv>=1.0.0

# ── Database — Supabase ───────────────────────────────────────
supabase>=2.4.0

# ── Streamlit UI ─────────────────────────────────────────────
streamlit>=1.35.0
streamlit-audiorec>=0.1.3        # push-to-talk audio recorder widget

# ── STT: Sarvam AI ───────────────────────────────────────────
requests>=2.31.0                 # Sarvam STT is REST-based

# ── TTS: ElevenLabs ──────────────────────────────────────────
elevenlabs>=1.2.0

# ── RAG (Bonus) ───────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=2.7.0     # local embeddings (no extra API key)

# ── Utilities ─────────────────────────────────────────────────
numpy>=1.26.0
